yourballot/issue/models/factories/issue_question.py
from factory.django import DjangoModelFactory

from yourballot.issue.models.issue_question import IssueQuestion


class IssueQuestionFactory(DjangoModelFactory):
    class Meta:
        model = IssueQuestion
