pyproject.toml
[tool.poetry]
name = "yourballot"
version = "0.1.0"
description = ""
authors = ["peteherman <20931612+peteherman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mypy = "^1.8.0"
django = "^5.0.3"
djangorestframework = "^3.14.0"
django-stubs = "^4.2.7"
factory-boy = "^3.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
django-stubs = "^4.2.7"
mypy = "^1.9.0"
djangorestframework-stubs = "^3.14.5"
isort = "^5.13.2"
types-factory-boy = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
